Remove ResourceSystemExtension
