--- Runtime/Core/Resource/ResourceSystemExtension.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 01b8abd13383b314ea752638c2e0cc76
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
